refactor: example imports and shared timestamp fn
