+++ test/unit/config/configTest.ts
@@ -281,7 +281,7 @@ describe('Config', () => {
       stubMethod($$.SANDBOX, ConfigFile.prototype, ConfigFile.prototype.read.name).callsFake(async function () {
         // @ts-expect-error -> this is any
         // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-        this.setContentsFromFileContents({ unknown: 'unknown config key and value' }, BigInt(Date.now()));
+        this.setContentsFromFileContents({ unknown: 'unknown config key and value' });
       });
 
       const config = await Config.create({ isGlobal: true });
