+++ test/unit/config/lwwMapTest.ts
@@ -6,16 +6,17 @@
  */
 import { expect, config } from 'chai';
 import { LWWMap, LWWState, SYMBOL_FOR_DELETED } from '../../../src/config/lwwMap';
+import { nowBigInt } from '../../../src/util/time';
 
 config.truncateThreshold = 0;
-const OLD_TIMESTAMP_OFFSET = BigInt(10_000_000_000_000);
+const TIMESTAMP_OFFSET = BigInt(1_000_000);
 
 describe('LWWMap', () => {
   type TestType = { foo: string; baz: string; opt?: number; optNull?: null };
   describe('all properties are known', () => {
     const state = {
-      foo: { value: 'bar', timestamp: process.hrtime.bigint() },
-      baz: { value: 'qux', timestamp: process.hrtime.bigint() },
+      foo: { value: 'bar', timestamp: nowBigInt() },
+      baz: { value: 'qux', timestamp: nowBigInt() },
     };
     let lwwMap: LWWMap<TestType>;
 
@@ -77,8 +78,8 @@ describe('LWWMap', () => {
       });
       it('all are updated', () => {
         const remoteState = {
-          foo: { value: 'bar2', timestamp: process.hrtime.bigint() },
-          baz: { value: 'qux2', timestamp: process.hrtime.bigint() },
+          foo: { value: 'bar2', timestamp: nowBigInt() },
+          baz: { value: 'qux2', timestamp: nowBigInt() },
         } satisfies LWWState<TestType>;
         lwwMap.merge(remoteState);
         expect(lwwMap.state).to.deep.equal(remoteState);
@@ -86,8 +87,8 @@ describe('LWWMap', () => {
       });
       it('all are deleted', () => {
         const remoteState = {
-          foo: { value: SYMBOL_FOR_DELETED, timestamp: process.hrtime.bigint() },
-          baz: { value: SYMBOL_FOR_DELETED, timestamp: process.hrtime.bigint() },
+          foo: { value: SYMBOL_FOR_DELETED, timestamp: nowBigInt() },
+          baz: { value: SYMBOL_FOR_DELETED, timestamp: nowBigInt() },
         } satisfies LWWState<TestType>;
         lwwMap.merge(remoteState);
         expect(lwwMap.state).to.deep.equal(remoteState);
@@ -95,8 +96,8 @@ describe('LWWMap', () => {
       });
       it('none are updated', () => {
         const remoteState = {
-          foo: { value: 'bar2', timestamp: process.hrtime.bigint() - OLD_TIMESTAMP_OFFSET },
-          baz: { value: 'qux2', timestamp: process.hrtime.bigint() - OLD_TIMESTAMP_OFFSET },
+          foo: { value: 'bar2', timestamp: nowBigInt() - TIMESTAMP_OFFSET },
+          baz: { value: 'qux2', timestamp: nowBigInt() - TIMESTAMP_OFFSET },
         } satisfies LWWState<TestType>;
         lwwMap.merge(remoteState);
         expect(lwwMap.state).to.deep.equal(state);
@@ -105,10 +106,10 @@ describe('LWWMap', () => {
       it('one is update, one is added, one is deleted', () => {
         lwwMap.set('opt', 4);
         const remoteState = {
-          foo: { value: 'bar2', timestamp: process.hrtime.bigint() },
-          baz: { value: 'qux2', timestamp: process.hrtime.bigint() - OLD_TIMESTAMP_OFFSET },
-          opt: { value: SYMBOL_FOR_DELETED, timestamp: process.hrtime.bigint() },
-          optNull: { value: null, timestamp: process.hrtime.bigint() },
+          foo: { value: 'bar2', timestamp: nowBigInt() },
+          baz: { value: 'qux2', timestamp: nowBigInt() - TIMESTAMP_OFFSET },
+          opt: { value: SYMBOL_FOR_DELETED, timestamp: nowBigInt() + TIMESTAMP_OFFSET },
+          optNull: { value: null, timestamp: nowBigInt() },
         } satisfies LWWState<TestType>;
         lwwMap.merge(remoteState);
         expect(lwwMap.get('foo')).to.equal('bar2');
@@ -122,8 +123,8 @@ describe('LWWMap', () => {
   describe('open-ended objects', () => {
     type OpenEndedType = TestType & { [key: string]: string };
     const initialState = {
-      foo: { value: 'bar', timestamp: process.hrtime.bigint() },
-      baz: { value: 'qux', timestamp: process.hrtime.bigint() },
+      foo: { value: 'bar', timestamp: nowBigInt() },
+      baz: { value: 'qux', timestamp: nowBigInt() },
     } satisfies LWWState<OpenEndedType>;
 
     it('set a new prop', () => {
@@ -155,9 +156,9 @@ describe('LWWMap', () => {
   });
   describe('nested objects', () => {
     const state = {
-      foo: { value: 'bar', timestamp: process.hrtime.bigint() },
-      baz: { value: 'qux', timestamp: process.hrtime.bigint() },
-      obj: { value: { a: 1, b: 2, c: 3 }, timestamp: process.hrtime.bigint() },
+      foo: { value: 'bar', timestamp: nowBigInt() },
+      baz: { value: 'qux', timestamp: nowBigInt() },
+      obj: { value: { a: 1, b: 2, c: 3 }, timestamp: nowBigInt() },
     };
     let lwwMap: LWWMap<{ foo: string; baz: string; opt?: number; optNull?: null }>;
