+++ typedocExamples/status/streamingClientExamples.ts
@@ -1,5 +1,5 @@
 import { ensureJsonMap, ensureString, JsonCollection, JsonMap } from '@salesforce/ts-types';
-import { Org } from '../../src/org';
+import { Org } from '../../src/org/org';
 import { StreamingClient, StatusResult } from '../../src/status/streamingClient';
 
 import { HttpRequest } from 'jsforce';
